Fix hardfault by switching to heap allocation for dhcp_server and dns_server
The hardfault issue was caused by stack-based memory allocation for dhcp_server and dns_server,
which led to dangling pointers and invalid memory. The memory was allocated on the stack during start_TCP_server
function and when the function returned the structures for DHCP and DNS servers got removed from the stack.
The fix involved switching to dynamic memory allocation using pvPortCalloc for both dhcp_server and dns_server.
This ensures the server structures will persist even after function returns, preventing the hardfault.

- Replaced stack allocation with heap allocation for dhcp_server.
- Replaced stack allocation with heap allocation for dns_server.